python/cyberres-mcp/pyproject.toml
#
# Copyright contributors to the agentic-ai-cyberres project
#
[project]
name = "cyberres-mcp"
version = "0.1.0"
description = "MCP server for validating recovered infrastructure resources including Linux VMs, Oracle databases, and MongoDB clusters. Provides tools to check network connectivity, OS health, database connectivity, and replica status with secure credential handling and structured logging."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.14",
    "fastmcp>=2.10.6",
    "mcp[cli]>=1.10.1",
    "oracledb>=3.3.0",
    "paramiko>=3.5.1",
    "pydantic>=2.11.7",
    "pymongo>=4.14.1",
    "requests>=2.32.4",
    "uvicorn>=0.35.0",
    "python-dotenv>=1.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
cyberres-mcp = "cyberres_mcp.server:main"
